拆分物品管理
